requirements.txt
celery[redis]==5.4.0
redis==5.0.4
Django==5.0.6
python-dotenv==1.0.1
requests
mssql-django
pywin32